cmake, rename, optz
